Implemented authorization for Edit and Delete question using Gate in both views and controllers

+++ app/Http/Controllers/QuestionController.php
@@ -65,7 +65,10 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
-       return view('questions.edit',compact('question'));
+        if(\Gate::denies('update-question',$question)){
+           abort(403, "Access Denied"); 
+        } 
+        return view('questions.edit',compact('question'));  
     }
 
     /**
@@ -77,6 +80,9 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question',$question)){
+            abort(403, "Access Denied"); 
+        } 
         $question->update($request->only('title','body'));
         return redirect('/questions')->with("success","Your question has been updated successfully");
     }
@@ -89,6 +95,9 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question',$question)){
+            abort(403, "Access Denied"); 
+        } 
         $question->delete();
         return redirect('/questions')->with('success','Your question has been deleted');
     }
